Overhauls and cleanups * Started a lot of work on migrating to floating point angles * Implemented double text box inputs * Removed some old code in Robot Arm * Disconnected the increments from the sliders, as now the floating point values of them is used to command the robot * Removed old camera detection * Removed old code for running camera in another thread * Did a lengthy rewrite of the RASM interpreter, cleaning up a lot. More cleanup will come in due time, for now old code is kept as comments in order not to lose it * Implemented a target system for programs * Changed program loading behavior to reflect new ISA and program binary format * Changed numeric variables to now check if more space is needed in the vector, this caused issues in the past * Explicitly clear the arrays in order to counteract a memory issue. Note: Issue still present.

+++ lib/libRobotExtension/RobotArm.hpp
@@ -37,7 +37,7 @@ class RobotArm : public ArmDevice
 
         //void servo_write6(float angles[6], uint16_t time);
 
-        std::vector<uint16_t> home_position;
+        std::vector<float32_t> home_position;
         
     friend class MainWindow;
 };
